EZEE-2303: Moved Validator to EzSystems\EzPlatformRichText\eZ\RichText\Validator namespace
