+++ src/bundle/Resources/config/fieldtype_services.yml
@@ -81,21 +81,21 @@ services:
         lazy: true
 
     ezrichtext.validator.input.ezxhtml5:
-        class: EzSystems\EzPlatformRichText\eZ\RichText\ValidatorAggregate
+        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator\ValidatorAggregate
         arguments: [!tagged ezrichtext.validator.input.ezxhtml5]
 
     ezrichtext.validator.docbook:
-        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator
+        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator\Validator
         arguments: ['%ezrichtext.validator.docbook.resources%']
         tags:
             - { name: 'ezrichtext.validator.input.ezxhtml5' }
 
     ezrichtext.validator.output.ezxhtml5:
-        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator
+        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator\Validator
         arguments: ['%ezrichtext.validator.output.ezxhtml5.resources%']
 
     ezrichtext.validator.input.dispatcher:
-        class: EzSystems\EzPlatformRichText\eZ\RichText\ValidatorDispatcher
+        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator\ValidatorDispatcher
         arguments:
             -
                 http://docbook.org/ns/docbook: null
@@ -103,7 +103,7 @@ services:
                 http://ez.no/namespaces/ezpublish5/xhtml5: '@ezrichtext.validator.output.ezxhtml5'
 
     ezrichtext.validator.internal_link:
-        class: EzSystems\EzPlatformRichText\eZ\RichText\InternalLinkValidator
+        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator\InternalLinkValidator
         arguments:
             - '@ezpublish.spi.persistence.cache.contentHandler'
             - '@ezpublish.spi.persistence.cache.locationHandler'
