+++ src/lib/eZ/RichText/Validator/CustomTagsValidator.php
@@ -6,8 +6,9 @@
  */
 declare(strict_types=1);
 
-namespace EzSystems\EzPlatformRichText\eZ\RichText;
+namespace EzSystems\EzPlatformRichText\eZ\RichText\Validator;
 
+use EzSystems\EzPlatformRichText\eZ\RichText\ValidatorInterface;
 use DOMDocument;
 use DOMXPath;
