+++ src/lib/eZ/settings/fieldtype_services.yml
@@ -17,23 +17,23 @@ services:
                 http://docbook.org/ns/docbook: null
 
     ezrichtext.validator.input.ezxhtml5:
-        class: EzSystems\EzPlatformRichText\eZ\RichText\ValidatorAggregate
+        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator\ValidatorAggregate
         arguments: [!tagged ezrichtext.validator.input.ezxhtml5]
 
     ezrichtext.validator.docbook:
-        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator
+        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator\Validator
         arguments: ['%ezrichtext.validator.docbook.resources%']
         tags:
             - { name: 'ezrichtext.validator.input.ezxhtml5' }
 
     ezrichtext.validator.input.dispatcher:
-        class: EzSystems\EzPlatformRichText\eZ\RichText\ValidatorDispatcher
+        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator\ValidatorDispatcher
         arguments:
             -
                 http://docbook.org/ns/docbook: null
 
     ezrichtext.validator.internal_link:
-        class: EzSystems\EzPlatformRichText\eZ\RichText\InternalLinkValidator
+        class: EzSystems\EzPlatformRichText\eZ\RichText\Validator\InternalLinkValidator
         arguments:
             - '@ezpublish.spi.persistence.cache.contentHandler'
             - '@ezpublish.spi.persistence.cache.locationHandler'
@@ -44,7 +44,7 @@ services:
         class: EzSystems\EzPlatformRichText\eZ\RichText\Normalizer\Aggregate
 
     # Symfony 3.4+ service definitions:
-    EzSystems\EzPlatformRichText\eZ\RichText\CustomTagsValidator:
+    EzSystems\EzPlatformRichText\eZ\RichText\Validator\CustomTagsValidator:
         public: false
         arguments: ['%ezplatform.ezrichtext.custom_tags%']
         tags:
